requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
pillow
numpy
faiss-cpu
transformers
jinja2
dghs-imgutils
torch
safetensors
